comisiones/requirements.txt
dash==2.17.1
dash-bootstrap-components==1.6.0
pandas==2.2.3
plotly==5.24.0
gunicorn==21.2.0
mysql-connector-python==9.0.0
numpy==1.26.4
sqlalchemy==2.0.31